+++ Cookid/Cookid/Repository/UserRepositrory.swift
@@ -43,49 +43,3 @@ class UserRepository {
     }
 }
 
-
-
-
-class UserService {
-    
-    func loadUserInfo(userID: String, completion: @escaping (User) -> Void) {
-        UserRepository.shared.fetchUserInfo(userID: userID) { userEntity in
-            var users: User!
-            
-            for userEntity in userEntity {
-                let userinfo = User(
-                    nickname: userEntity.nickname,
-                    determination: userEntity.determination,
-                    priceGoal: userEntity.priceGoal,
-                    userType: UserType(rawValue: userEntity.userType)!)
-                users = userinfo
-            }
-            completion(users)
-        }
-    }
-}
-
-
-class MealService {
-    
-    static let shared = MealService()
-    
-    func loadMeal(completion: @escaping (Meal) -> Void) {
-        MealRepository.shared.fetchMeals { mealEntity in
-            var meals: Meal!
-            
-            for mealEntity in mealEntity {
-
-                let meal = Meal(price: mealEntity.price,
-                                date: mealEntity.date.StringTodate()!,
-                                name: mealEntity.name,
-                                image: "box",
-                                mealType: MealType(rawValue: mealEntity.mealType)!,
-                                mealTime: MealTime(rawValue: mealEntity.mealTime)!)
-                meals = meal
-            }
-            completion(meals)
-        }
-    }
-}
-
